add library

+++ local_docker.sh
@@ -2,10 +2,18 @@
 
 cd ./infra/docker
 
-docker-compose build
-
-if [ ! -e ../node_modules ]; then
-  docker-compose run --rm web sh -c "npm install"
+if [ $1 == "build" ]; then
+  docker-compose build
 fi
 
-docker-compose up -d
+if [ $1 == "up" ]; then
+  if [ ! -e ../node_modules ]; then
+    docker-compose run --rm web sh -c "npm install"
+  fi
+
+  docker-compose up -d
+fi
+
+if [ $1 == "down" ]; then
+  docker-compose stop
+fi
